fix all bugs and warnings

+++ FileCabinetApp/Handlers/RemoveCommandHandler.cs
@@ -8,22 +8,22 @@ namespace FileCabinetApp.Handlers
         public RemoveCommandHandler(IFileCabinetService service) : base(service)
         {
         }
-        
-        public override void SetNext(ICommandHandler handler)
-        {
-            throw new NotImplementedException();
-        }
 
+        /// <summary>
+        /// Remove record from current service. If service storage is file, it marks the record as deleted
+        /// </summary>
+        /// <param name="request">Object contains command and it's parameters</param>
+        /// <exception cref="ArgumentNullException">Request in null</exception>
         public override void Handle(AppCommandRequest request)
         {
-            throw new NotImplementedException();
-        }
-        
-        private void Remove(string parameters)
-        {
+            if (request is null)
+            {
+                throw new ArgumentNullException(nameof(request));
+            }
+            
             try
             {
-                var id = Convert.ToInt32(parameters, CultureInfo.InvariantCulture);
+                var id = Convert.ToInt32(request.Parameters, CultureInfo.InvariantCulture);
                 
                 Service.Remove(id);
